requirements.txt
langchain>=0.0.335
python-dotenv>=1.0.0
streamlit>=1.25.0
watchdog>=3.0.0
openai
